chore: rename to match consistency
